Fix accelerometer offset calculation in calibrateMPU9250 method
Calibration code was not correctly preserving the LSB of the accelerometer offset register (used for temperature correction), and was not subtracting the full value from the accel offset. BSL the measured average bias before subtracting from register value produces the desired result. Also cleaned up some copypasta in the comments.